Rename `auth/client` to `auth/shared`
