+++ pkg/cmd/auth/shared/client.go
@@ -1,4 +1,4 @@
-package client
+package shared
 
 import (
 	"fmt"
